chore: adjust pixel and social proof
